% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtdownload.R
\name{download}
\alias{download}
\title{Download files from CGHub - a wrapper around gtdownload}
\usage{
download(uuids, path = getwd(), gtdownload_exe = "gtdownload", cores = 2,
  key = get_opts("cghub_key_file"), params = "")
}
\arguments{
\item{uuids}{}

\item{path}{}

\item{gtdownload_exe}{}

\item{cores}{}

\item{key}{}

\item{params}{Additional parameters to be supplied}
}
\description{
gt.cgDownload
}
\examples{
\dontrun{

download("89324e86-5b7a-4f69-92c1-3b67293f8748")

}
}

