% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gtdownload.R
\name{gt.cgDownload}
\alias{gt.cgDownload}
\title{gt.cgDownload}
\usage{
gt.cgDownload(uuids, path = "~/cgdump", gtdownload_exe = "gtdownload",
  cores = 2, key = "mykey.pem", params = "")
}
\arguments{
\item{uuids}{}

\item{path}{}

\item{gtdownload_exe}{}

\item{cores}{}

\item{key}{}

\item{params}{Additional parameters to be supplied}
}
\description{
gt.cgDownload
}

