% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgquery.R
\name{gt.cgquery}
\alias{gt.cgquery}
\title{gt.cgquery}
\usage{
gt.cgquery(search, url = "https://cghub.ucsc.edu/cghub/metadata/analysisFull",
  count.only, cores = 1)
}
\arguments{
\item{search}{}

\item{url}{}

\item{count.only}{}

\item{cores}{}
}
\description{
gt.cgquery
}
\examples{
\dontrun{
search = 'library_strategy=WXS&state=live&disease_abbr=CESC'
gt.cgquery(search = search, count.only = TRUE, cores = 5)
}
}

